Enabling translation via custom zod schemas
- now pharmacist schema errors became translated

+++ src/controllers/Pharmacist/create.controller.ts
@@ -8,19 +8,22 @@ const createPharmacist = async (req: Request, res: TypedResponse<PharmacistRespo
     try {
         const validatedData: CreatePharmacistDto = req.validatedData;
         const processed: string[] = [];
-        for (const file of req.files as Express.Multer.File[]) {
-            const processedImage = await processImage(file, {
-                userId: req.params.id,
-                supportsWebP: res.locals.supportsWebP,
-                isLegacyBrowser: res.locals.isLegacyBrowser,
-            });
-            processed.push(processedImage.imageURL);
-            try {
-                await fs.unlink(file.path);
-            } catch (e) {
-                console.log(e);
+        if (req.files) {
+            for (const file of req.files as Express.Multer.File[]) {
+                const processedImage = await processImage(file, {
+                    userId: req.params.id,
+                    supportsWebP: res.locals.supportsWebP,
+                    isLegacyBrowser: res.locals.isLegacyBrowser,
+                });
+                processed.push(processedImage.imageURL);
+                try {
+                    await fs.unlink(file.path);
+                } catch (e) {
+                    console.log(e);
+                }
             }
         }
+
         const doc = await Pharmacist.create({
             ...validatedData,
             syndicateRecords: [
@@ -30,6 +33,11 @@ const createPharmacist = async (req: Request, res: TypedResponse<PharmacistRespo
                     registrationNumber: validatedData.registrationNumber,
                 },
             ],
+            currentSyndicate: {
+                syndicate: "نقابة الصيادلة المركزية",
+                startDate: validatedData.registrationDate,
+                registrationNumber: validatedData.registrationNumber,
+            },
             images: processed,
         });
         const pharmacist = await handlePharmacistFields(doc);
