+++ src/controllers/Pharmacist/licenses/updateLicense.controller.ts
@@ -47,19 +47,21 @@ const updateLicense = async (req: Request, res: TypedResponse<PharmacistResponse
             imagesURLs = oldImages;
         }
         // handle uploaded images
-        for (const file of req.files as Express.Multer.File[]) {
-            const processedImage = await processImage(file, {
-                userId: req.params.id,
-                supportsWebP: res.locals.supportsWebP,
-                isLegacyBrowser: res.locals.isLegacyBrowser,
-            });
-            if (!imagesURLs.includes(processedImage.imageURL)) {
-                imagesURLs.push(processedImage.imageURL);
-            }
-            try {
-                await fs.unlink(file.path);
-            } catch (e) {
-                console.log(e);
+        if (req.files) {
+            for (const file of req.files as Express.Multer.File[]) {
+                const processedImage = await processImage(file, {
+                    userId: req.params.id,
+                    supportsWebP: res.locals.supportsWebP,
+                    isLegacyBrowser: res.locals.isLegacyBrowser,
+                });
+                if (!imagesURLs.includes(processedImage.imageURL)) {
+                    imagesURLs.push(processedImage.imageURL);
+                }
+                try {
+                    await fs.unlink(file.path);
+                } catch (e) {
+                    console.log(e);
+                }
             }
         }
